% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countRegional.R
\name{countRegional}
\alias{countRegional}
\title{Sumamry of prescribing volume for a regional unit}
\usage{
countRegional(data, region, datename = datename, unit = Quantity)
}
\arguments{
\item{data}{A dataframe with date and unit of medicines}

\item{region}{Column name of regional unit of interest (e.g., `practiceid`, `gp_cluster`)}

\item{datename}{Name of the colume containing the date data}

\item{unit}{Unit of interest (e.g., `quantity`, `items`)}
}
\value{
Dataframe of units prescribed grouped by date and regional unit
}
\description{
Function to create tsibble object for a count of `unit` per `datename` e.g. countRegional(oc, Locality, datename, Quantity) is the `Quantity` per `datename` grouped by `Locality` from `oc`. `Quantity` is corrected for the number of days in each month.
}
\examples{
\dontrun{countRegional(data, region = practiceid)}

}
