% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fableModels.R
\name{fableModels}
\alias{fableModels}
\title{Fable model fits}
\usage{
fableModels(data, value = n)
}
\arguments{
\item{data}{A dataframe with time as `datename` and quantity as `n`}

\item{value}{Response variable}
}
\value{
Dataframe of fable model fits
}
\description{
Fit various fable time series models.
}
\examples{
\dontrun{fableModels(data)}
}
