% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fableModelsBound.R
\name{fableModelsBound}
\alias{fableModelsBound}
\title{Fable model fits for bound data}
\usage{
fableModelsBound(data)
}
\arguments{
\item{data}{A dataframe with time as `datename` and quantity as `n`}
}
\value{
Dataframe of fable model fits
}
\description{
Fit various fable time series models.
}
\examples{
\dontrun{fableModelsBound(data)}
}
