% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glsData.R
\name{glsData}
\alias{glsData}
\title{Prepare data for GLS analysis}
\usage{
glsData(data)
}
\arguments{
\item{data}{A drug to filter `differences` by}
}
\value{
Dataframe
}
\description{
Function to prepare difference data for GLS analysis
}
\examples{
\dontrun{glsData(drug)}

}
