% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monthlyPrescribing.R
\name{monthlyPrescribing}
\alias{monthlyPrescribing}
\title{Prepare dispensing data for time series analysis}
\usage{
monthlyPrescribing(data)
}
\arguments{
\item{data}{Data to be prepared}
}
\value{
tsibble
}
\description{
Function to prepare dispensing data for time series analysis
}
\examples{
\dontrun{monthlyPrescribing(drug)}

}
