% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{my_scaled_logit}
\alias{my_scaled_logit}
\title{my_scaled_logit}
\usage{
my_scaled_logit(x, lower = 0, upper = 100)
}
\arguments{
\item{x}{Double to transform}

\item{lower}{Lower limit of admissible values}

\item{upper}{Upper limit of admissible values}
}
\description{
Transforming data to be bound between a lower and upper limit http://www.maths.bristol.ac.uk/R/web/packages/fable/vignettes/transformations.html
}
