% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prophetModels.r
\name{prophetModels}
\alias{prophetModels}
\title{Prophet model fits}
\usage{
prophetModels(data)
}
\arguments{
\item{data}{A dataframe with time as `datename` and quantity as `n`}
}
\value{
Dataframe of prophet model fits
}
\description{
Fit various prophet time series models.
}
\examples{
\dontrun{prophetModels(data)}

}
