% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prophetModelsBound.R
\name{prophetModelsBound}
\alias{prophetModelsBound}
\title{Prophet model fits for bound data}
\usage{
prophetModelsBound(data)
}
\arguments{
\item{data}{A dataframe with time as `datename` and quantity as `n`}
}
\value{
Dataframe of prophet model fits
}
\description{
Fit various prophet time series models.
}
\examples{
\dontrun{prophetModelsBound(data)}

}
