% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/propRegional.R
\name{propRegional}
\alias{propRegional}
\title{Summary of proportion of a medicine precribed for a regional unit over time}
\usage{
propRegional(data, region, datename = datename, unit = Quantity, bnf)
}
\arguments{
\item{data}{A dataframe with date and unit of medicines with multiple BNF codes}

\item{region}{Column name of regional unit of interest (e.g., `practiceid`, `gp_cluster`)}

\item{datename}{Name of the column containing the date data}

\item{unit}{Unit of interest (e.g., `quantity`, `items`)}

\item{bnf}{BNF code of medicine of interest}
}
\value{
Dataframe of proportion of `bnf` prescribed grouped by date and regional unit
}
\description{
Function to create tsibble object for a count of `unit` per `datename` e.g. countRegional(oc, Locality, datename, Quantity) is the `Quantity` per `datename` grouped by `Locality` from `oc`. `Quantity` is corrected for the number of days in each month.
}
\examples{
\dontrun{propRegional(data, region = practiceid)}

}
