% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regionalCV.r
\name{regionalCV}
\alias{regionalCV}
\title{Cross-validated model fits}
\usage{
regionalCV(data, cv_dist = 6, init = 36, step = 1)
}
\arguments{
\item{data}{A dataframe with time as `datename` and quantity as `n`.}

\item{cv_dist}{The number of months that should be cross-validated.}

\item{init}{The number of months that cross-validations should start from.}

\item{step}{The number of months between each cross-validation step.}
}
\value{
Dataframe of prophet model fits
}
\description{
Cross-validated fits of various fable and prophet models. Will take advantage of plan() if called outside of function.
}
\examples{
\dontrun{regionalCV(data)}

}
