% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regionalJointFcstsBound.R
\name{regionalJointFcstsBound}
\alias{regionalJointFcstsBound}
\title{Bound forecasts for each regional unit}
\usage{
regionalJointFcstsBound(data, h = "8 months")
}
\arguments{
\item{data}{A dataframe with time as `datename` and quantity as `n`}

\item{h}{The number of months that should forecast}
}
\value{
Model forecasts
}
\description{
Various fable forecasts for monthly quantity for each regional_unit. Can make use of plan() futures.
}
\examples{
\dontrun{regionalJointFcsts(data, h ="6 months")}

}
