% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regionalJointFits.r
\name{regionalJointFits}
\alias{regionalJointFits}
\title{Forecasts for each regional unit}
\usage{
regionalJointFits(data)
}
\arguments{
\item{data}{A dataframe with time as `datename` and quantity as `n`.}
}
\value{
Dataframe of model fits
}
\description{
Various fable and prophet forecasts for monthly quantity for each regional_unit. Will take advantage of plan() if called outside of function.
}
\examples{
\dontrun{regionalJointFcsts(data)}

}
