% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{scaled_logit}
\alias{scaled_logit}
\title{Helper functions}
\usage{
scaled_logit(x, lower = 0, upper = 1)
}
\arguments{
\item{x}{Double to transform}

\item{lower}{Lower limit of admissible values}

\item{upper}{Upper limit of admissible values

Transforming data to be bound between a lower and upper limit http://www.maths.bristol.ac.uk/R/web/packages/fable/vignettes/transformations.html}
}
\description{
Helper functions for serCymruTools
}
