% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fableModelsBound.R
\name{fableModelsBound}
\alias{fableModelsBound}
\title{Fable model fits}
\usage{
fableModelsBound(data, lower = 0, upper = 100)
}
\arguments{
\item{data}{A dataframe with time as `datename` and quantity as `n`}

\item{lower}{Lower limit of admissible values}

\item{upper}{Upper limit of admissible values}
}
\value{
Dataframe of fable model fits
}
\description{
Fit various fable time series models.
}
\examples{
\dontrun{fableModelsBound(data, lower = 0, upper = 100)}
}
