% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectBestModel.R
\name{rankModels}
\alias{rankModels}
\title{rankModels}
\usage{
rankModels(data)
}
\arguments{
\item{data}{Tibble of model accuracies including rmse and winkler}
}
\description{
Rank models according to rmse and winkler scores
}
