% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regionalJointFcsts.r
\name{regionalJointFcsts}
\alias{regionalJointFcsts}
\title{Forecasts for each regional unit}
\usage{
regionalJointFcsts(data, h = "6 months")
}
\arguments{
\item{data}{A dataframe with time as `datename` and quantity as `n`.}

\item{h}{The number of months that should forecast.}
}
\value{
Model forecasts
}
\description{
Various fable and prophet forecasts for monthly quantity for each regional_unit. Can make use of plan() futures.
}
\examples{
\dontrun{regionalJointFcsts(data)}

}
